Fixes and improvements for Snowflake (#370)
* Fix Snowflake constructor taking an Instant (subtract milliseconds before shifting bits to the left).

* Fix SnowflakeMark.elapsedNow returning Duration with wrong sign (before this, Snowflake.min.timeMark.hasPassedNow() returned false and Snowflake.max.timeMark.hasPassedNow() returned true).

* Fix Snowflake.timeMark passing unadjusted milliseconds since Discord Epoch to SnowflakeMark's constructor by using same math as Snowflake.timeStamp.
Example where this comes up: Snowflake(Clock.System.now()).timeMark.elapsedNow() should return a very short duration, but did not before.

* Deduplicate timeStamp creation logic in Snowflake.timeStamp and Snowflake.timeMark.

* Using ushr instead of shr in Snowflake to make timestamp related functionality behave correctly for Long values smaller than 0 (having 1 as the first bit).
Maybe ULong should be used for Snowflake.value? (Would bring more accurate Snowflake.max with ULong.MAX_VALUE, avoid confusion with negative Snowflake values etc. but I'm not sure about breaking changes and serialization.)

* Additional documentation for compareTo and equals behavior for Snowflake.

* Snowflake: tests, checks in constructor that takes instant, endOfTime constant, more documentation

* Fix test naming

* Fix test naming again

* Use ULong for Snowflake value

* Change Snowflake.Serializer from class to object

* Add pseudo Snowflake constructor that takes a Long value

* Don't use java.math.BigInteger for Test

* Remove unnecessary range for Random.nextULong()

+++ core/src/test/kotlin/rest/RestTest.kt
@@ -41,7 +41,7 @@ fun imageBinary(path: String): Image {
 @EnabledIfEnvironmentVariable(named = "KORD_TEST_TOKEN", matches = ".+")
 class RestServiceTest {
 
-    private val publicGuildId = Snowflake(322850917248663552)
+    private val publicGuildId = Snowflake(322850917248663552u)
 
     private val token = System.getenv("KORD_TEST_TOKEN")
 
@@ -387,7 +387,7 @@ class RestServiceTest {
     @Order(21)
     fun `errors are thrown correctly`() = runBlocking {
         val exception = assertThrows<RestRequestException> {
-            runBlocking { kord.getChannel(Snowflake(-500)) }
+            runBlocking { kord.getChannel(Snowflake((-500).toULong())) }
         }
 
         assert(exception.error != null)
